Add some logging

+++ src/music/music_manager.hpp
@@ -13,7 +13,7 @@
 
 class MusicManager {
 public:
-    using ID = uint64_t;
+    using ID = uint32_t;
     using Playlist = Playlist<ID>;
 
 private:
